cardgame/requirements.txt
django-debug-toolbar
Django>=3.0.3,<4.0.0
psycopg2>=2.8.4,<2.9.0 # for postgres
djangorestframework>=3.11.0,<3.12.0